### 2019-04-24 #1411 @whedon
[REVIEW]: “Blaster” – a graphical user interface for common sequences analysis tools
**Submitting author:** @snowformatics (<a href="http://orcid.org/0000-0003-0536-835X">Stefanie Lueck</a>)
**Repository:** <a href="https://github.com/snowformatics/BlasterQt" target ="_blank">https://github.com/snowformatics/BlasterQt</a>
**Version:** v.1.2.4

- [x] **Version:** Does the release version given match the GitHub release (v.1.2.4)?
