### 2019-06-04 #1411 @vbarrera
Hi @kyleniemeyer.
Sorry for the radio silence.
I've been having some issues with some parts of the app. I'll be adding my notes this week.
