### 2019-06-09 snowformatics/BlasterQt#9 @vbarrera
Databases names ignore everything after "."
Hi.

When creating new databases if the name includes ".", everything after is ignored.

For example:
HS.GRh38 becomes HS.

I'd recommend either to indicate it in the documentation or parse it somehow.

Best.
