### 2019-06-09 snowformatics/BlasterQt#10 @vbarrera
Editing database name in the Database management section doesn't have any effect
Hi.

Editing and changing the name in the Database management section doesn't have any effect. Although the name changes in that section, the other tools doesn't seem to detect the change.

Best.
