### 2019-06-09 snowformatics/BlasterQt#11 @vbarrera
Missing documentation
Hi.

Although the help section is quite helpful, some tools lack any information. For example, for the format converter tools there's no information on how to use the multi/single FASTA converters. It isn't clear how to organize the files and folders to obtain the desired result.

Best.
