### 2019-06-09 snowformatics/BlasterQt#12 @vbarrera
Need for remove file option
Hi.

When using the tool, if you provide a file for any of the section, it can't be removed. If used afterwards using the text field, it continues using the file rather than the text file so the results don't change.

Best.
