### 2019-06-09 #1411 @vbarrera
Initial submitted issues:
https://github.com/snowformatics/BlasterQt/issues/12
https://github.com/snowformatics/BlasterQt/issues/11
https://github.com/snowformatics/BlasterQt/issues/10
https://github.com/snowformatics/BlasterQt/issues/9
https://github.com/snowformatics/BlasterQt/issues/4
