### 2019-06-20 #1411 @lpantano
Hi @snowformatics, just wondering if you have seen the reviewer's comment and you have any question on your side.
