### 2019-06-20 #1411 @snowformatics
Hi Lorena,

I apologize for my silence and I thank Victor for his very constructive and
helpful review, I think all points could be addressed and fixed. I have a
question, what is the deadline for fixing the bugs? I am currently moving
with baby and it's a bit complicated. Of course I will try ASAP to submit a
new version but I am not done yet.

Thanks for your understanding
Stefanie

Lorena Pantano <notifications@github.com> schrieb am Do., 20. Juni 2019,
17:32:

> Hi @snowformatics <https://github.com/snowformatics>, just wondering if
> you have seen the reviewer's comment and you have any question on your side.
>
