### 2019-08-04 #1411 @labarba
@snowformatics — Could we have a status update on your revisions? I see you were moving back in June, but hopefully things have settled for you. If not, and you will need extended time to work on your submission, it would help if you let us know and we can add the "paused" label here.
