### 2019-08-05 #1411 @snowformatics
Dear Lorena,

I apologize for my silence. I am working on the revision but the movement
took much longer then expected.  Since yesterday we are settled and now I
am going to work more often on the software.  A paused label would be very
helpful,  I put myself a deadline for end of August.

Thanks for your understanding and tolerance

Stefanie

Am So., 4. Aug. 2019, 13:04 hat Lorena A. Barba <notifications@github.com>
geschrieben:

> @snowformatics <https://github.com/snowformatics> — Could we have a
> status update on your revisions? I see you were moving back in June, but
> hopefully things have settled for you. If not, and you will need extended
> time to work on your submission, it would help if you let us know and we
> can add the "paused" label here.
>
