### 2019-08-23 4dbbeeb
Issue #9 fixed
Spaces and interpuncts will be automatically replaced with '_'.

A note was added to the wizard page.
