### 2019-08-23 snowformatics/BlasterQt#9 @snowformatics
Fixed with
https://github.com/snowformatics/BlasterQt/commit/4dbbeebb990147d1951ba7f4d15d6254856b9f94
