### 2019-08-23 01a61ad
enable to edit cells in table widget of the database
