### 2019-08-23 snowformatics/BlasterQt#10 @snowformatics
Fixed with
https://github.com/snowformatics/BlasterQt/commit/01a61ad2602ae5fceea6a3424bf6eb4390fa907b

Actually editing database name should be enabled
