### 2019-08-24 snowformatics/BlasterQt#11 @snowformatics
The documentation including screenshots were updated.
http://labtools.ipk-gatersleben.de/help/BlasterQt/BlasterQt.pdf
