### 2019-08-27 #1411 @snowformatics
Dear reviewers,

I fixed and closed all issues except of two (snowformatics/BlasterQt#12
<https://github.com/snowformatics/BlasterQt/issues/12> and
snowformatics/BlasterQt#4
<https://github.com/snowformatics/BlasterQt/issues/4>) because I can not
reproduce the bugs.

@vbarrera <https://github.com/vbarrera>  would it be possible to sent me
the sequences which causes the problems?

Thank you very much in advance and again sorry for the late reply

Cheers

Am Mo., 5. Aug. 2019 um 19:27 Uhr schrieb whedon <notifications@github.com>:

> Reminder set for @snowformatics <https://github.com/snowformatics> in 26
> days
>
