### 2019-09-03 #1411 @vbarrera
@snowformatics Is the binary updated? I wanted to try the new features and confirm the other issues with the most updated one. Thanks.

Also @labarba and @lpantano, I'll be away for 3 weeks starting next week. Should be paused the review?
