### 2019-09-08 #1411 @snowformatics
The binary is not updated yet, I will do it next week.

I have no problem with pausing the review.

Am Mi., 4. Sep. 2019, 00:45 hat Victor Barrera <notifications@github.com>
geschrieben:

> @snowformatics <https://github.com/snowformatics> Is the binary updated?
> I wanted to try the new features and confirm the other issues with the most
> updated one. Thanks.
>
> Also @labarba <https://github.com/labarba> and @lpantano
> <https://github.com/lpantano>, I'll be away for 3 weeks starting next
> week. Should be paused the review?
>
