### 2019-10-18 #1411 @lpantano
@vbarrera , @snowformatics , it is ok to unpause this and continue with the review?
