### 2019-10-19 #1411 @snowformatics
I am fine with unpausing, should I create a new binary nevertheless of 2
open issues?

Am Fr., 18. Okt. 2019, 13:45 hat Lorena Pantano <notifications@github.com>
geschrieben:

> @vbarrera <https://github.com/vbarrera> , @snowformatics
> <https://github.com/snowformatics> , it is ok to unpause this and
> continue with the review?
>
