### 2019-10-26 #1411 @vbarrera
Hi.
I'm good with unpausing too.

@snowformatics It would be helpful if you create the new binary with the updated version.
I'll reevaluate all issues and close them if they're no longer happening.
Thank you.
