### 2019-11-01 #1411 @snowformatics
Hi,

I created a new binary:

https://sourceforge.net/projects/blasterqt/files/BlasterQtv1.3.exe/download

Best regards
Stefanie
