### 2019-11-03 snowformatics/BlasterQt#10 @vbarrera
No longer possible to change database name.
Is that the desired behavior? Although not vital, changing the database name might be useful.
