### 2019-11-03 snowformatics/BlasterQt#12 @vbarrera
Closing this issue as a new version is out.
