### 2019-11-03 snowformatics/BlasterQt#13 @vbarrera
Fasta Extractor: sequences names issues
When using the FASTA extractor utility the following problems are encountered:
- ">" is not ignored and if ">Sequence1" is used, sequence is not found
- Same issue with white spaces at the end, as ">Sequence1 "
- If multiple sequence IDs are used, only the the results for the first one are provided.

Attaching screenshots.

Best,
![Sequence_extract](https://user-images.githubusercontent.com/14740681/68086983-71e84700-fe1f-11e9-9fd1-72a0fda2ead5.png)
![Sequence_extract_2](https://user-images.githubusercontent.com/14740681/68086984-71e84700-fe1f-11e9-8eb9-1fe232955c0d.png)
![Sequence_extract_3](https://user-images.githubusercontent.com/14740681/68086985-71e84700-fe1f-11e9-8084-f38adf2476b7.png)
![Sequence_extract_4](https://user-images.githubusercontent.com/14740681/68086986-71e84700-fe1f-11e9-99c7-d4341964ffea.png)
![Sequence_extract_5](https://user-images.githubusercontent.com/14740681/68086987-7280dd80-fe1f-11e9-9e08-1d25697ed077.png)
