### 2019-11-03 snowformatics/BlasterQt#14 @vbarrera
Clear fields options doesn't work on some menus
The clear fields option doesn't work on all menus.

For example:
- In Multiple Alignment, nothing is removed.
- Sequence converter, only the text field is cleaned. No way of removing the file field. This causes extra problem as when running the analysis, the file takes priority over the text box.
- Format converter. Idem as sequence extractor.

Best,

Screenshots attached:

![clear_fields_1](https://user-images.githubusercontent.com/14740681/68087232-c12f7700-fe21-11e9-9fdd-9a6205f64f8f.png)
![clear_fields_2](https://user-images.githubusercontent.com/14740681/68087233-c12f7700-fe21-11e9-9e8a-06450dde4ba7.png)
![clear_fields_3](https://user-images.githubusercontent.com/14740681/68087234-c12f7700-fe21-11e9-8e47-06937f5c5cc3.png)
![clear_fields_4](https://user-images.githubusercontent.com/14740681/68087235-c12f7700-fe21-11e9-87b2-db63060dd315.png)
![clear_fields_5](https://user-images.githubusercontent.com/14740681/68087231-c096e080-fe21-11e9-8ca2-e57cb02ca9d4.png)
