### 2019-11-03 snowformatics/BlasterQt#15 @vbarrera
Files with extension ".fa" are not identified as valid
Improvement:
Files with ".fa" extension are not recognized by the open file (or choose source for database creation) menus.

Although this can be solved by changing the name, some important databases (like ensembl) provide data files with that extension as default.
