### 2019-11-03 snowformatics/BlasterQt#16 @vbarrera
Uninstaller leaves some files behind
Improvement:
For cleaner uninstall.
When using the uninstaller "unins000.exe", some files are left in C:\Users\User\AppData\Local\BlasterQt

Best,
