### 2019-11-03 snowformatics/BlasterQt#11 @vbarrera
Thanks.
Updates would be also needed to the help menu.
![help_menu](https://user-images.githubusercontent.com/14740681/68088441-5126ee00-fe2d-11e9-82f8-68a32640e5c3.png)
Best
