### 2019-11-03 snowformatics/BlasterQt#17 @vbarrera
Test materials
Improvement:

When using the test materials:

test_blastn.txt doesn't generate same results as indicated:
https://github.com/snowformatics/BlasterQt/blob/master/test/output_blastn.txt

Attached result:

![test_blastn](https://user-images.githubusercontent.com/14740681/68088546-43be3380-fe2e-11e9-9b3e-77e39cdf6060.png)

This is not critical, but test should be updated when the rest of the issues have been solved.

Best.
