### 2019-11-03 snowformatics/BlasterQt#18 @vbarrera
csv to multi fasta not working properly
Using the csv to multi fasta doesn't work correctly.

Attached images, file used and log file.

[tmpubunnz.zip](https://github.com/snowformatics/BlasterQt/files/3801794/tmpubunnz.zip)

![csvMultiFasta](https://user-images.githubusercontent.com/14740681/68088606-cfd05b00-fe2e-11e9-909b-3658b1f2f06e.png)
![csvMultiFasta_2](https://user-images.githubusercontent.com/14740681/68088605-cfd05b00-fe2e-11e9-9e54-5d1c6e1e9796.png)

[blaster_main.exe.log](https://github.com/snowformatics/BlasterQt/files/3801797/blaster_main.exe.log)
