### 2019-11-03 snowformatics/BlasterQt#19 @vbarrera
Multifasta file to Single Fasta Files, no headers
Headers are lost from multifasta file to single fasta files.

[test_multi.zip](https://github.com/snowformatics/BlasterQt/files/3801801/test_multi.zip)

Best.
