### 2019-11-03 #1411 @vbarrera
Hi,
Thank you very much for creating the new binary.

I've been testing the new version and I've found the following issues:
- Added a comment to the DB database management:
https://github.com/snowformatics/BlasterQt/issues/10#issuecomment-549142471
- Comment on Documentation:
https://github.com/snowformatics/BlasterQt/issues/11#issuecomment-549154709
- https://github.com/snowformatics/BlasterQt/issues/13
- https://github.com/snowformatics/BlasterQt/issues/14
-  https://github.com/snowformatics/BlasterQt/issues/15
- https://github.com/snowformatics/BlasterQt/issues/16
- https://github.com/snowformatics/BlasterQt/issues/17
- https://github.com/snowformatics/BlasterQt/issues/18
- https://github.com/snowformatics/BlasterQt/issues/19

Best regards
