### 2019-11-12 snowformatics/BlasterQt#19 @snowformatics
Fixed in BlasterQtv1.4
