### 2019-11-12 snowformatics/BlasterQt#14 @snowformatics
Fixed in BlasterQtv1.4
