### 2019-11-12 snowformatics/BlasterQt#15 @snowformatics
Good point added in BlasterQtv1.4
