### 2019-11-12 2ef315d
Update output_blastn.txt
https://github.com/snowformatics/BlasterQt/issues/17
