### 2019-11-12 snowformatics/BlasterQt#17 @snowformatics
Updated in BlasterQtv1.4
