### 2019-11-12 snowformatics/BlasterQt#19 @snowformatics
Fixed with BlasterQtv1.4
