### 2019-11-12 snowformatics/BlasterQt#13 @snowformatics
Fixed with BlasterQtv1.4
