### 2019-11-12 snowformatics/BlasterQt#18 @snowformatics
Fixed in BlasterQtv1.4
