### 2019-11-12 snowformatics/BlasterQt#16 @snowformatics
Thanks for your comment. We decided to leave the databases and the log files to easily install new versions without to create again the databases. This was a request from several users.

However it's not a problem to create an uninstaller which deletes the database files for the publication (deleting all files is not recommended by InnoSetup see http://www.jrsoftware.org/ishelp/index.php?topic=uninstalldeletesection.

Just let me know what you prefer.
