### 2019-11-12 snowformatics/BlasterQt#10 @snowformatics
We discussed this issue and some user prefer to enabled this function because they reported a renaming by mistake. Hope that's ok, otherwise let me know.
