### 2019-11-12 snowformatics/BlasterQt#11 @snowformatics
Of course, done.
