### 2019-11-12 #1411 @snowformatics
Hi,

thank you very much for extensively testing, reporting bugs and improvements.
I created an new binary of Version 1.4 :
https://sourceforge.net/projects/blasterqt/files/BlasterQtv1.4.exe/download

Best regards
Stefanie
