### 2019-11-12 7515709
Update paper.md
Help and download files updated
