### 2019-12-08 snowformatics/BlasterQt#16 @vbarrera
I agree with the explanation.
Thanks.
