### 2019-12-08 #1411 @vbarrera
@snowformatics
Thanks for solving all the issues.
Software is working properly and all issues have been addressed.

Please, for the final approval, update documents to describe the latest version:
- testing_guidelines (i.e: FASTQ to FASTA is described and it's no longer a functionality of the software).
- update README to include those changes.
- update manual to include changes. Also, non English word detected (i.e: Inhaltsverzeichnis). (minor edit).
- All texts/links make reference to v.1.2.4
- Although references are included in the text, they don't appear in the reference section of the paper (minor edit).

Best regards,
