+++ README.md
@@ -46,13 +46,27 @@ All kind of contributions (documentation, testing, adding more features, suggest
 
 # History
 Blaster v. 1.3
+
 Removed FASTQ conversation feature
+
 Removed editing database names
+
 Updated documentation
+
 Bug fixes
+
 "." or spaces in database names will be replaced with "_" (see https://github.com/snowformatics/BlasterQt/issues/9)
 
+
 Blaster v. 1.4
+
+Updated documentation
+
+Bug fixes
+
+Added ".fa" extension 
+
+Updated test material
 
 
 
