### 2019-12-09 #1411 @snowformatics
@vbarrera

Thank you very much for reviewing the manuscript and your patience, I think the software got major improvements.
I updated the documents and fixed the reference issue.

Best regards
Stefanie
