### 2019-12-15 #1411 @lpantano
Thank you! @snowformatics could you make a Zenodo archive from the current github version. The tittle and authors should match the paper exactly. Let me know if you have any question.
