### 2019-12-15 #1411 @snowformatics
Hi Lorena,

I have created an archive with DOI at our institute repository system, is
this ok for you? Otherwise I could create a Zenodo archive, no problem.

http://dx.doi.org/10.5447/IPK/2019/23

Am So., 15. Dez. 2019 um 17:09 Uhr schrieb Lorena Pantano <
notifications@github.com>:

> Thank you! @snowformatics <https://github.com/snowformatics> could you
> make a Zenodo archive from the current github version. The tittle and
> authors should match the paper exactly. Let me know if you have any
> question.
>
