### 2019-12-15 #1411 @lpantano
Thanks for the quick reply. Sadly, the zenodo needs to contains all the code in this repo. You can link zenodo with the GitHub code, and every time you do a release it would create an archive version. When you do the link you need to modify the title an author but that's all. Or you can repeat what you did here, but it needs to be the full repository. I hope this is ok with you. Sorry to not have explained better before.
