### 2019-12-17 #1411 @lpantano
@openjournals/joss-eics, I think this is ready for you.
