### 2019-12-18 #1411 @arfon
I'm going to leave this issue open until the PDF is rendering properly on the JOSS site (this looks to be a GitHub pages issue)
